Respect helm name overrides for sercrets
The secret names are not aligned with rest of the manifest generated by
helm when `fullnameOverride` or `nameOverride` is used in a helm chart.
This is at least confusing. The expected name of secrets can be changed by setting
the release name during `helm install`, but this might misalign with the release names
of other applications

This change is backward incompatible for deployments where
`fullnameOverride` or `nameOverride` was used, because it changes the name of the secrets.